requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.7
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2023.2
sqlparse==0.4.3
